move: fetch 유틸리티 함수 위치 변경
